+++ include/glaze/core/to.hpp
@@ -1,0 +1,32 @@
+// Glaze Library
+// For the license information refer to glaze.hpp
+
+#pragma once
+
+#include "glaze/core/common.hpp"
+#include "glaze/core/opts.hpp"
+
+// Common behavior for `to` specializations, typically applies for all formats
+
+namespace glz::detail
+{
+   template <uint32_t Format>
+   struct to<Format, hidden>
+   {
+      template <auto Opts>
+      static void op(auto&& value, auto&&...) noexcept
+      {
+         static_assert(false_v<decltype(value)>, "hidden type should not be written");
+      }
+   };
+   
+   template <uint32_t Format>
+   struct to<Format, skip>
+   {
+      template <auto Opts>
+      static void op(auto&& value, auto&&...) noexcept
+      {
+         static_assert(false_v<decltype(value)>, "skip type should not be written");
+      }
+   };
+}
